Texture provider framework, textures work

+++ src/runtime/render/public/material.h
@@ -39,7 +39,7 @@ public:
 	{}
 
 	EBlendMode _blend;
-	uint16 _baseTexture;
+	uint32 _baseTexture;
 
 	uint64 getHash() const;
 };
